requirements.txt
matplotlib==3.8.3
seaborn==0.13.1
streamlit==1.35.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
tensorflow==2.15.0
tensorflow-hub==0.16.1
opencv-python==4.10.0.84